+++ frontend/src/modules/customerCRUD.js
@@ -79,14 +79,16 @@ const GetCustomers = () => {
 
   // DELETE CUSTOMER BY ID
   const DeleteCustomer = (_id) => {
-    fetch(baseURL + "/customers/delete/" + _id, {
-      method: "DELETE",
-      headers: {
-        "auth-token": localStorage.getItem("Token"),
-      },
-    }).then(() => {
-      GetUsersCustomers(); // Updates page
-    });
+    return confirm("Are you sure you want to delete this Customer?").then(
+      fetch(baseURL + "/customers/delete/" + _id, {
+        method: "DELETE",
+        headers: {
+          "auth-token": localStorage.getItem("Token"),
+        },
+      }).then(() => {
+        GetUsersCustomers(); // Updates page
+      })
+    );
   };
 
   // UPDATE CUSTOMER BY ID
