+++ frontend/src/modules/tenantCRUD.js
@@ -95,14 +95,16 @@ const GetTenants = () => {
 
   // DELETE BY ID
   const DeleteTenant = (_id) => {
-    fetch(baseURL + "/tenants/delete/" + _id, {
-      method: "DELETE",
-      headers: {
-        "auth-token": localStorage.getItem("Token"),
-      },
-    }).then(() => {
-      GetUsersTenants(); // Updates page
-    });
+    return confirm("Are you sure you want to delete this Tenant?").then(
+      fetch(baseURL + "/tenants/delete/" + _id, {
+        method: "DELETE",
+        headers: {
+          "auth-token": localStorage.getItem("Token"),
+        },
+      }).then(() => {
+        GetUsersTenants(); // Updates page
+      })
+    );
   };
 
   // UPDATE BY ID
